bc command implemented in c language.
This project aims to develop a bc command implemented by myself in C language. 
The software shall be able to operate in unlimited length arbitrary precision floating points. It can carry out all arithmetic operations and may also be used for evaluating a few selected operations. 
Any suggestions for any scope of improvement in the code are most welcome.